requirements.txt
matplotlib==3.10.3
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
numpy==1.26.4
pillow==10.4.0
safetensors==0.5.3
tqdm==4.67.1
transformers==4.52.4
tokenizers==0.21.1